+++ components/checkbox/Group.jsx
@@ -16,10 +16,14 @@ export default React.createClass({
     onChange: React.PropTypes.func,
   },
   getInitialState() {
-    const { value, defaultValue } = this.props;
-    return {
-      value: value || defaultValue,
-    };
+    const props = this.props;
+    let value;
+    if ('value' in props) {
+      value = props.value;
+    } else if ('defaultValue' in props) {
+      value = props.defaultValue;
+    }
+    return { value };
   },
   componentWillReceiveProps(nextProps) {
     if ('value' in nextProps) {
@@ -30,7 +34,7 @@ export default React.createClass({
   },
   toggleOption(option) {
     const optionIndex = this.state.value.indexOf(option);
-    const value = this.state.value;
+    const value = [...this.state.value];
     if (optionIndex === - 1) {
       value.push(option);
     } else {