+++ components/transfer/list.jsx
@@ -69,8 +69,8 @@ class TransferList extends Component {
   }
 
   render() {
-    const { prefixCls, dataSource, titleText, filter, checkedKeys,
-            checkStatus, body, footer, showSearch } = this.props;
+    const { prefixCls, dataSource, titleText, filter, checkedKeys, notFoundContent,
+            checkStatus, body, footer, showSearch, searchPlaceholder } = this.props;
 
     // Custom Layout
     const footerDom = footer({ ...this.props });
@@ -81,21 +81,19 @@ class TransferList extends Component {
       [`${prefixCls}-with-footer`]: !!footerDom,
     });
 
-    const showItems = dataSource.map((item) => {
-      // apply filter
+    const showItems = dataSource.filter((item) => {
       const itemText = this.props.render(item);
       const filterResult = this.matchFilter(itemText, filter);
+      return !!filterResult;
+    }).map((item) => {
       const renderedText = this.props.render(item);
-
-      if (filterResult) {
-        return (
-          <li onClick={this.handleSelect.bind(this, item)} key={item.key} title={renderedText}>
-            <Checkbox checked={checkedKeys.some(key => key === item.key)} />
-            {renderedText}
-          </li>
-        );
-      }
-    }).filter(item => !!item);
+      return (
+        <li onClick={this.handleSelect.bind(this, item)} key={item.key} title={renderedText}>
+          <Checkbox checked={checkedKeys.some(key => key === item.key)} />
+          {renderedText}
+        </li>
+      );
+    });
 
     return (
       <div className={listCls} {...this.props}>
@@ -104,19 +102,23 @@ class TransferList extends Component {
             prefixCls: 'ant-transfer',
             checked: checkStatus === 'all',
             checkPart: checkStatus === 'part',
-            checkable: <span className={`ant-transfer-checkbox-inner`}></span>
+            checkable: <span className={'ant-transfer-checkbox-inner'}></span>
           })}<span className={`${prefixCls}-header-selected`}><span>{(checkedKeys.length > 0 ? `${checkedKeys.length}/` : '') + dataSource.length} 条</span>
           <span className={`${prefixCls}-header-title`}>{titleText}</span></span>
         </div>
         { bodyDom ||
         <div className={ showSearch ? `${prefixCls}-body ${prefixCls}-body-with-search` : `${prefixCls}-body`}>
           { showSearch ? <div className={`${prefixCls}-body-search-wrapper`}>
-            <Search prefixCls={`${prefixCls}-search`} onChange={this.handleFilter.bind(this)} handleClear={this.handleClear.bind(this)} value={filter} />
+            <Search prefixCls={`${prefixCls}-search`}
+              onChange={this.handleFilter.bind(this)}
+              handleClear={this.handleClear.bind(this)}
+              placeholder={searchPlaceholder}
+              value={filter} />
           </div> : null }
           <Animate component="ul"
             transitionName={this.state.mounted ? `${prefixCls}-highlight` : ''}
             transitionLeave={false}>
-            {showItems.length > 0 ? showItems : <div className={`${prefixCls}-body-not-found`}>Not Found</div>}
+            {showItems.length > 0 ? showItems : <div className={`${prefixCls}-body-not-found`}>{notFoundContent}</div>}
           </Animate>
         </div>}
         { footerDom ? <div className={`${prefixCls}-footer`}>